requirements.txt
absl-py
opencv-python
numpy